+++ src/rviz_legged_plugins/rviz_legged_plugins-extras.cmake
@@ -1,0 +1,3 @@
+# find package Qt5 because otherwise using the rviz_default_plugins::rviz_default_plugins
+# exported target will complain that the Qt5::Widgets target does not exist
+find_package(Qt5 REQUIRED QUIET COMPONENTS Widgets)